Add test, with some reorganization as reczones is not linked into testrunner.
So move a few functions to reczones-helpers.cc
